Added some comments to make the code cleaner. Removed unused WeatherCache.
