--- Rainy/Model/Caching/WeatherCache.swift
@@ -1,17 +1,0 @@
-//
-//  WeatherCache.swift
-//  Rainy
-//
-//  Created by Andrea Gottardo on 2020-04-08.
-//  Copyright © 2020 Andrea Gottardo. All rights reserved.
-//
-
-import Foundation
-
-class WeatherCache {
-    static let shared = WeatherCache()
-
-    private init() {}
-
-    func weather(for _: Location) {}
-}
