--- Rainy/Model/Weather Data/TempUnit.swift
@@ -1,14 +1,0 @@
-//
-//  TempUnit.swift
-//  Rainy
-//
-//  Created by Andrea Gottardo on 14/08/2019.
-//  Copyright © 2019 Andrea Gottardo. All rights reserved.
-//
-
-import Foundation
-
-public enum TempUnit: Int {
-    case celsius
-    case fahrenheit
-}
