+++ Rainy/Model/Weather Data/WeatherUpdate.swift
@@ -6,12 +6,10 @@
 //  Copyright © 2017 Andrea Gottardo. All rights reserved.
 //
 
-import CoreLocation
 import Foundation
 import UIKit
 
-// MARK: - WeatherUpdate
-
+/// A Swift Codable representing the response returned by the DarkSky.net API.
 struct WeatherUpdate: Codable {
     let latitude, longitude: Double?
     let timezone: String?
@@ -22,88 +20,7 @@ struct WeatherUpdate: Codable {
     let alerts: [WeatherAlert]?
 }
 
-class Temperature: Codable {
-    private let doubleValue: Double
-
-    func stringRepresentation(settingsManager: SettingsManager = .shared) -> String {
-        switch settingsManager.temperatureUnit {
-        case .celsius:
-            return String(lround(0.55555 * (doubleValue - 32))) + "°"
-        case .fahrenheit:
-            return String(lround(doubleValue)) + "°"
-        }
-    }
-
-    init(_ double: Double) {
-        doubleValue = double
-    }
-
-    required init(from: Decoder) {
-        do {
-            let container = try from.singleValueContainer()
-            doubleValue = try container.decode(Double.self)
-        } catch {
-            Log.error("Failed to parse temperature, not a double.")
-            doubleValue = 0
-        }
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(doubleValue)
-    }
-}
-
-enum WeatherIcon: String, Codable {
-    case clearDay = "clear-day"
-    case clearNight = "clear-night"
-    case rain
-    case snow
-    case sleet
-    case wind
-    case fog
-    case cloudy
-    case partlyCloudyDay = "partly-cloudy-day"
-    case partlyCloudyNight = "partly-cloudy-night"
-    case hail
-    case thunderstorm
-    case tornado
-
-    var icon: UIImage? {
-        let configuration: UIImage.SymbolConfiguration = .init(weight: .regular)
-        switch self {
-        case .clearDay:
-            return UIImage(systemName: "sun.max", withConfiguration: configuration)
-        case .clearNight:
-            return UIImage(systemName: "moon.stars", withConfiguration: configuration)
-        case .rain:
-            return UIImage(systemName: "cloud.rain", withConfiguration: configuration)
-        case .snow:
-            return UIImage(systemName: "cloud.snow", withConfiguration: configuration)
-        case .sleet:
-            return UIImage(systemName: "cloud.sleet", withConfiguration: configuration)
-        case .wind:
-            return UIImage(systemName: "wind", withConfiguration: configuration)
-        case .fog:
-            return UIImage(systemName: "cloud.fog", withConfiguration: configuration)
-        case .cloudy:
-            return UIImage(systemName: "cloud", withConfiguration: configuration)
-        case .partlyCloudyDay:
-            return UIImage(systemName: "cloud.sun", withConfiguration: configuration)
-        case .partlyCloudyNight:
-            return UIImage(systemName: "cloud.moon", withConfiguration: configuration)
-        case .hail:
-            return UIImage(systemName: "cloud.hail", withConfiguration: configuration)
-        case .tornado:
-            return UIImage(systemName: "tornado", withConfiguration: configuration)
-        case .thunderstorm:
-            return UIImage(systemName: "cloud.bolt", withConfiguration: configuration)
-        }
-    }
-}
-
-// MARK: - Currently
-
+/// Represents the weather conditions for a 1-hour interval starting at `time`.
 struct HourCondition: Codable {
     let time: Int?
     let summary: String?
@@ -120,16 +37,15 @@ struct HourCondition: Codable {
     let precipType: String?
 }
 
-// MARK: - Daily
-
+/// Represents the weather conditions for multiple days. The number of days returned (`==data.count`)
+/// depends on the location and data availability.
 struct DailyData: Codable {
     let summary: String?
     let icon: String?
     let data: [DailyCondition]?
 }
 
-// MARK: - Datum
-
+/// Represents the weather conditions for a 24-hours interval.
 struct DailyCondition: Codable {
     let time: Int?
     let summary: String?
@@ -163,8 +79,8 @@ struct DailyCondition: Codable {
     let apparentTemperatureMaxTime: Int?
 }
 
-// MARK: - Hourly
-
+/// Represents the weather conditions for multiple hours. The number of hours returned (`==data.count`)
+/// depends on the location and data availability.
 struct HourlyData: Codable {
     let summary: String?
     let icon: WeatherIcon?
